requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
httptools==0.1.1
idna==2.10
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
multidict==5.1.0
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.5
numpy==1.20.2
pandas==1.2.4
Pillow==8.1.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
rake-nltk==1.0.4
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
sanic==21.3.2
sanic-routing==0.4.3
scikit-surprise==1.1.1
scipy==1.6.2
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
uvloop==0.15.2
websockets==8.1
yarl==1.6.3
